Create storage admin page and streamline pricing copy
