+++ apps/web/app/admin/storage/page.tsx
@@ -1,0 +1,7 @@
+import ClientPage from "./ClientPage";
+
+export const metadata = { title: "Admin – Storage | Pineapple Tapped" };
+
+export default function Page() {
+  return <ClientPage />;
+}
